Split tarball progress test from write_test.go
This example test reaches out to the internet, which isn't ideal.

For now, I'm just going to split it off into a separate file so it's
easier to exclude.

Also, fix some whitespace nits.
